% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\title{Prepare a schema diagram of relations between data.frames and/or csv/tsv files}
\usage{
schema(
  data_frames = list(),
  data_frame_names = character(0),
  csv_files = character(0),
  output_type = "grViz",
  all_links = FALSE
)
}
\arguments{
\item{data_frames}{A list of data.frames.}

\item{data_frame_names}{A character vector of names of data.frame objects.}

\item{csv_files}{A character vector of file names, file paths, or urls pointing
to csv/tsv files. This value is passed to \code{\link[data.table]{fread}}'s
argument/parameter \code{input}.}

\item{output_type}{A string -- one of the three options: \code{"grViz"} (default) or
\code{"gv"} or \code{"svg"}. See the return value description below.}

\item{all_links}{Logical (Boolean) -- if \code{FALSE} (default) all linkages
between the data.frames are included in a graph (as edges); if \code{TRUE} only the
minimum necessary linkages (edges) are included and some linkages can be
inferred indirectly as passing through "intermediary" data.frames.}
}
\value{
One of the three possibilities:
\itemize{
 \item if \code{output_type} is \code{"grViz"}: a diagram plot produced by
       \code{\link[DiagrammeR]{grViz}} (object of class \code{htmlwidget})
 \item if \code{output_type} is \code{"gv"}: a string (character vector of length 1)
       with the GraphViz code (which can be later edited for tweaks)
 \item if \code{output_type} is \code{"svg"}: a string (character vector of length 1)
       with the svg (xml) code produced by \code{\link[DiagrammeRsvg]{export_svg}}
}
}
\description{
This function uses the column names and column types/classes of data.frames
and/or csv files to produce a schema diagram of relations between these
tables. By default, "correct" relations (same column names and same types/classes)
are drawn as solid black lines. If there are evident type/class mismatches
despite the same column names, the relation is drawn as a striped red line.
If there is a slight mismatch that can be overcome by automatic coercion
(integer-double, but both columns numeric) the relation line is dashed but black. \cr \cr
At least one data.frame or csv/tsv file needs to be specified. In other words,
either \code{`data_frames`} or \code{`data_frame_names`} or \code{`csv_files`} parameter
needs to be specified. It is allowed to use all the three arguments/parameters
simultaneously i.e. have some data.frames provided directly (as values),
some others indirectly (as object names) and yet some others as csv/tsv
file names/paths/urls to be imported via \code{\link[data.table]{fread}}.
}
